function.object vector and double purge

+++ test/std/utilities/function.objects/func.memfn/member_data.pass.cpp
@@ -16,7 +16,7 @@
 
 struct A
 {
-    double data_;
+    long long data_;
 };
 
 template <class F>
